+++ src/bin.ts
@@ -9,7 +9,7 @@ import * as path from 'node:path';
 import type { Options } from 'tiged';
 import { tiged } from 'tiged';
 import glob from 'tiny-glob/sync.js';
-import { base, pathExists, tryRequire } from './utils';
+import { base, pathExists, tryRequire } from './utils.js';
 
 const args = mri<Options & { help?: string }>(process.argv.slice(2), {
   alias: {