+++ tsup.config.ts
@@ -2,52 +2,54 @@ import * as path from 'node:path';
 import type { Options } from 'tsup';
 import { defineConfig } from 'tsup';
 
-export default defineConfig(options => {
-  const commonOptions: Options = {
+const tsupConfig = defineConfig((overrideOptions): Options[] => {
+  const commonOptions = {
     clean: true,
-    format: ['cjs', 'esm'],
-    tsconfig: path.resolve('tsconfig.build.json'),
-    sourcemap: true,
-    shims: true,
-    splitting: false,
     removeNodeProtocol: false,
-    ...options,
-  };
+    shims: true,
+    sourcemap: true,
+    splitting: false,
+    target: ['esnext', 'node20'],
+    tsconfig: path.resolve('tsconfig.build.json'),
+    ...overrideOptions,
+  } satisfies Options;
 
   return [
     {
       ...commonOptions,
       name: 'Modern ESM',
+      entry: { index: 'src/index.ts' },
       format: ['esm'],
-      entry: ['src/index.ts'],
     },
     {
       ...commonOptions,
       name: 'CJS Development',
+      entry: { index: 'src/index.ts' },
       format: ['cjs'],
-      entry: ['src/index.ts'],
     },
     {
       ...commonOptions,
       name: 'CLI Development',
-      entry: ['src/bin.ts'],
+      entry: { bin: 'src/bin.ts' },
       external: ['tiged'],
-      treeshake: 'smallest',
+      format: ['cjs', 'esm'],
       minify: true,
     },
     {
       ...commonOptions,
       name: 'ESM Type Definitions',
       dts: { only: true },
+      entry: { index: 'src/index.ts' },
       format: ['esm'],
-      entry: ['src/index.ts'],
     },
     {
       ...commonOptions,
       name: 'CJS Type Definitions',
-      format: ['cjs'],
       dts: { only: true },
-      entry: ['src/index.ts'],
+      entry: { index: 'src/index.ts' },
+      format: ['cjs'],
     },
   ];
 });
+
+export default tsupConfig;