+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.0.0
 
+- build: migrate to ESM ([#118](https://github.com/tiged/tiged/pull/118))
+
 - chore: fix `prettier` config ([#117](https://github.com/tiged/tiged/pull/117))
 
 - ci: Switch to Node.js v22 ([#115](https://github.com/tiged/tiged/pull/115))